requirements.txt
Django==2.2
pkg-resources==0.0.0
python-magic==0.4.15
pytz==2018.9
sqlparse==0.3.0
